+++ packages/ncform-theme-elementui/src/components/control-comps/textarea.vue
@@ -5,9 +5,10 @@
     v-show="!hidden"
     :placeholder="placeholder"
     type="textarea"
-    v-model="modelVal"
+    v-model="inputVal"
     :rows="mergeConfig.rows"
     :autosize="mergeConfig.autoSize"
+    @blur="onBlur"
   ></el-input>
 </template>
 
@@ -24,15 +25,21 @@
 
     mixins: [controlMixin],
 
+    created() {
+      this.$data.inputVal = this.$data.modelVal;
+    },
+
     props: {
       value: {
         type: String,
-        default: ''
+        default: '',
+        updateOn: 'change', // change or blur
       }
     },
 
     data() {
       return {
+        inputVal: '',
         // 组件特有的配置属性
         defaultConfig: {
           rows: 2,
@@ -42,12 +49,23 @@
       }
     },
 
+    watch: {
+      inputVal(newVal, oldVal) {
+        if ((!newVal && !oldVal) || this.mergeConfig.updateOn === 'blur') return;
+        this.$data.modelVal = newVal;
+      }
+    },
+
     computed: {
       // disabled / readonly / hidden / placeholder 你可以直接使用这些变量来控制组件的行为
     },
 
     methods: {
-
+      onBlur() {
+        if (this.mergeConfig.updateOn === 'blur') {
+          this.$data.modelVal = this.$data.inputVal;
+        }
+      },
     }
 
   }