+++ packages/ncform/src/components/vue-ncform/form-item.vue
@@ -6,7 +6,7 @@
     <component v-if="isNormalObjSchema(schema)" :is="'ncform-' + schema.ui.widget" :schema="schema" :form-data="formData" :temp-data="tempData" :global-const="globalConfig.constants" :idx-chain="idxChain" :config="schema.ui.widgetConfig">
 
       <template v-for="(fieldSchema, fieldName) in schema.properties" :slot="fieldName">
-        <form-item :schema="fieldSchema" :form-data="formData" :temp-data="tempData" :global-const="globalConfig.constants" :key="fieldName" :global-config="globalConfig" :complete-schema="completeSchema"></form-item>
+        <form-item :schema="fieldSchema" :form-data="formData" :temp-data="tempData" :global-const="globalConfig.constants" :key="fieldName" :global-config="globalConfig" :idx-chain="idxChain" :complete-schema="completeSchema" :paths="paths ? paths + '.' + fieldName : fieldName" :form-name="formName"></form-item>
       </template>
 
     </component>
@@ -15,7 +15,7 @@
     <component v-else-if="isNormalArrSchema(schema)" :is="'ncform-' + schema.ui.widget" :schema="schema" :form-data="formData" :temp-data="tempData" :global-const="globalConfig.constants" :idx-chain="idxChain" :config="schema.ui.widgetConfig" class="__ncform-control">
 
       <template v-for="(fieldSchema, fieldName) in (schema.items.properties || {__notObjItem: schema.items})" :slot="fieldName" slot-scope="props">
-        <form-item :schema="props.schema" :key="fieldName" :form-data="formData" :temp-data="tempData" :global-const="globalConfig.constants" :idx-chain="(idxChain ? idxChain + ',' : '') + props.idx" :global-config="globalConfig" :complete-schema="completeSchema"></form-item>
+        <form-item :schema="props.schema" :key="fieldName" :form-data="formData" :temp-data="tempData" :global-const="globalConfig.constants" :idx-chain="(idxChain ? idxChain + ',' : '') + props.idx" :global-config="globalConfig" :complete-schema="completeSchema" :paths="paths + '[' + props.idx + ']'" :form-name="formName"></form-item>
       </template>
 
     </component>
@@ -107,6 +107,7 @@ import "./layout-comps";
 import "./control-comps";
 import ncformCommon from '@ncform/ncform-common';
 import _get from "lodash-es/get";
+import _isArray from "lodash-es/isArray";
 
 const ncformUtils = ncformCommon.ncformUtils;
 
@@ -137,6 +138,13 @@ export default {
     idxChain: {
       // 用于记录在数组中的索引
       type: String
+    },
+    paths: {
+      // 用于记录字段的路径
+      type: String
+    },
+    formName: {
+      type: String
     }
   },
 
@@ -164,15 +172,6 @@ export default {
           constData: this.globalConfig.constants
         }
       });
-
-      if (this.$options._init4valueTemplate) { // Prevent init value from being overwritten
-        if (this.schema.value) result = this.schema.value;
-        // User nextTick will cause the init value to be incorrect when field item in list
-        // so here use setTimeout instead
-        setTimeout(() => {
-          this.$options._init4valueTemplate = false;
-        }, 100)
-      }
 
       return result;
     },
@@ -245,9 +244,24 @@ export default {
   },
 
   watch: {
-    valueTemplate(newVal) {
-      if (newVal !== undefined)
-        this.schema.value = newVal;
+    valueTemplate: {
+      handler: function(newVal, oldVal) {
+        if (newVal !== undefined) {
+          if (oldVal === undefined || JSON.stringify(newVal) !== JSON.stringify(oldVal)) {
+            if (!(this.$options._init4valueTemplate && (_isArray(this.schema.value) ? this.schema.value.length > 0 : this.schema.value))) { // Prevent init value from being overwritten
+              this.schema.value = newVal;
+            }
+            if (this.$options._init4valueTemplate) {
+              // User nextTick will cause the init value to be incorrect when field item in list
+              // so here use setTimeout instead
+              setTimeout(() => {
+                this.$options._init4valueTemplate = false;
+              }, 10)
+            }
+          }
+        }
+      },
+      immediate: true
     },
     schema: {
       handler: function(newVal) {
@@ -263,6 +277,16 @@ export default {
 
         // 对比控件改变前后的值，判断是否需要对其进行校验。
         if (changed) {
+          if (!ncformUtils.isNormalObjSchema(newVal) && !ncformUtils.isNormalArrSchema(newVal)) { // 叶子结点
+            const formVM = window.__$ncform.__ncFormsGlobalList[this.formName];
+            formVM.$emit('change', {
+              paths: this.paths,
+              itemValue: this.schema.value,
+              formValue: this.formData,
+              itemOldValue: this.$data.itemValue
+            })
+          }
+
           if (ncformUtils.isNormalArrSchema(newVal)) {
             this.$data.itemValue = ncformUtils.getModelFromSchema(newVal);
           } else {